requirements_elite.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
pyautogui==0.9.54
mss==9.0.1
Pillow>=10.3.0
openai>=1.6.1
anthropic>=0.21.3
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0
pywin32>=306; sys_platform == 'win32'
pywebview>=6.0